requirements.txt
openai==1.30.1
black==24.4.2
colorlog==6.8.2
pre-commit==3.7.0
transformers==4.41.0
requests==2.32.2
accelerate==0.30.1
fastapi==0.111.0
uvicorn==0.29.0
Pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
sentencepiece==0.1.99